flask-redis/requirements.txt
flask
redis
pytest
pytest-mock
requests-mock